# Conflicts:
#	done/Makefile
#	done/shell.c

+++ done/bmblock.c
@@ -104,7 +104,7 @@ void bm_print(struct bmblock_array *bmblock_array) {
     printf("max: %" PRIu64 "\n", bmblock_array->max);
     printf("cursor: %" PRIu64 "\n", bmblock_array->cursor);
     puts("content:");
-    for (int i = 0; i < bmblock_array->length; i++) {
+    for (uint64_t i = 0; i < bmblock_array->length; i++) {
         printf("%d:", i);
         print_binary(bmblock_array->bm[i]);
     }
